restructuring the python package
Changes to be committed:
	modified:   python_packages/profiles_classifier_pkg/profiles_classifier/model.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/Connector.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/Connector.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/License.md -> python_packages/profiles_classifier_pkg/profiles_classifier/src/License.md
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/MLTrainer.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/MLTrainer.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/README.md -> python_packages/profiles_classifier_pkg/profiles_classifier/src/README.md
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/RedshiftConnector.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/RedshiftConnector.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/SnowflakeConnector.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/SnowflakeConnector.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/__init__.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/__init__.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/conda_env.yml -> python_packages/profiles_classifier_pkg/profiles_classifier/src/conda_env.yml
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/config/model_configs.yaml -> python_packages/profiles_classifier_pkg/profiles_classifier/src/config/model_configs.yaml
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/constants.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/constants.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/logger.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/logger.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/predict.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/predict.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/requirements.txt -> python_packages/profiles_classifier_pkg/profiles_classifier/src/requirements.txt
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/tests/test_utils.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/tests/test_utils.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/train.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/train.py
	renamed:    python_packages/profiles_classifier_pkg/profiles_classifier/training_pkg/utils.py -> python_packages/profiles_classifier_pkg/profiles_classifier/src/utils.py
