+++ python_packages/profiles_classifier_pkg/profiles_classifier/src/SnowflakeConnector.py
